fix: always update on dev

+++ entrypoint.sh
@@ -1,8 +1,14 @@
 #!/bin/sh
 set -e
 
-php artisan migrate:refresh
-php artisan passport:install
-php artisan db:seed
+if [ -z "$FRESH_INSTALL" ]; then
+    echo "The variable \$FRESH_INSTALL is empty, using default init"
+    php artisan migrate
+else
+    php artisan migrate:refresh
+    php artisan passport:install --force
+    php artisan db:seed
+fi
+
 apache2-foreground
 # php-fpm
